% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{checkColTypes}
\alias{checkColTypes}
\title{Checks/converts the column types of an imported tibble to the selected column types}
\usage{
checkColTypes(data, coltypes = 0)
}
\arguments{
\item{data}{tibble to have columns checked}

\item{coltypes}{column types that are currently valid}
}
\value{
Data with the column types checked. If coltypes was the first row of the input data, this will now be removed.
}
\description{
The tibble can have the first row as the expected column types (removed from the output) or they can be given separately
Int = convert all to integer
String = convert to character
Name(s) = convert to character
Real = convert to numeric
Boolean = convert to logical and check if conversion was valid
(Long,Lat) = check that the coordinates are valid
YYYY-MM-DD or yyyy-mm-dd = check that year month and day are valid; no conversion
yyyy-mm-dd HH:MM:SS or yyyy-mm-dd hh:mm:ss = check that year month day h m s are valid; no conversion
DD.MM.YY HH:MM" or dd.mm.yy hh:mm = check that year month day h m s are valid and convert with to datetime as yyyy-mm-dd hh:mm:ss with lubridate
}
\examples{
projects = read_excel(path, sheet = "Projects"); projects = checkColTypes(projects)
}
