% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{combineFiles}
\alias{combineFiles}
\title{Function for combining files into a single csv}
\usage{
combineFiles(path, filepattern, fileformat, sheetchoice = "Daten", save = "n")
}
\arguments{
\item{path}{The folder containing the data to be combined}

\item{filepattern}{A string pattern that must be contained in the names of files to be combined}

\item{fileformat}{The format of the files to be combined}

\item{sheetchoice}{For excel files: the name of the sheet to be used}

\item{save}{The name of the combined file to be saved (default = "n" = don't save)}
}
\description{
Specify a folder, a format and a pattern, and all files of the specified format with the
pattern in their name in the chosen folder will be combined. The name of the output combined
file is given under the "save" input; if this is not include the data will be combined but not
automatically saved.
Note: all files to be combined MUST have the same column names!
}
