% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{dbSafeNames}
\alias{dbSafeNames}
\title{Function for creating "safe" names for postgres}
\usage{
dbSafeNames(names, uniq = FALSE)
}
\arguments{
\item{names}{names to be fixed; eg. from colnames(dataframe)}

\item{uniq}{if the names should all be unique set to true and _1, _2 etc. will be added (default = false)}
}
\description{
Removes protected characters and puts everything into lower case
}
