% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{editValuesInDatabase}
\alias{editValuesInDatabase}
\title{Function for writing updating line(s) of a table already in the database}
\usage{
editValuesInDatabase(values, tableName, key, keyvalues = NA)
}
\arguments{
\item{values}{updated data to go into table}

\item{tableName}{name of table in the database}

\item{key}{column used to compare the old and new data; new data will be updated where key matches keyvalue(s)}

\item{keyvalues}{value(s) used to compare the old and new data; do not set this parameter if you wish to
check all lines (can be slow!). More than one value can be set eg. c(key1,key2,...keyn).}
}
\description{
This function will: i) connect to database, ii) find lines matching keyvalue\link{1} in the designated key
column of the specified table and replace these line(s) of data with the line matching
keyvalue\link{1} in the new data table, iii) repeat for keyvalues\link{n}, iv) exit connection
}
\keyword{database}
