% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{FixIDs}
\alias{FixIDs}
\title{Fix all columns identified as "id" into db safenames (these columns are often used as keys)}
\usage{
FixIDs(data, uniq = FALSE)
}
\arguments{
\item{data}{data tibble to be checked and edited}

\item{uniq}{if the names should all be unique set to true and _1, _2 etc. will be added (default = false)}
}
\description{
This function finds all columns with "id" in their names, as these are expected to be primary
or foreign keys. dbSafeNames is applied to all IDs, as keys must conform to PostgreSQL requirements
for correct matching. Columns that have been fixed are listed in a print out.
}
