% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{queryDatabase}
\alias{queryDatabase}
\title{Function to query the postgres database from R}
\usage{
queryDatabase(query)
}
\arguments{
\item{query}{needs to be in postgres code}
}
\value{
The output of the query from postgres is returned; there will be no return if the query did not generate an output
}
\description{
This function sends a query written in PostgreSQL to the Functional Ecology database. The username
and password must be present in the workspace.
}
\examples{
queryDatabase("SELECT datetime, precip_tot FROM z_micromet WHERE datetime BETWEEN '01/01/2009 00:00' AND '01/01/2010 00:00'")
}
