% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{writeNewDataset}
\alias{writeNewDataset}
\title{Function for writing a new dataset and associated metadata file(s) from a csv to the database}
\usage{
writeNewDataset(
  data_path,
  metadata_path,
  datasetid,
  base_path = "/Volumes/daten_inst/oekophysiologie/",
  othermeta = 0
)
}
\arguments{
\item{data_path}{Path from the base path data folder to the csv data file}

\item{metadata_path}{Path from the base path data folder to the csv main metadata data file}

\item{datasetid}{ID for the dataset to be added (the dataset must already have been added to the datasets table!)}

\item{base_path}{Path from root to the data folder (normally set as the server: "/Volumes/daten_inst/oekophysiologie/" but you must be connected to the server for this to work)}

\item{othermeta}{Path from the base path data folder to extra csv metadata data files (optional)}
}
\description{
This function will: i) connect to database, ii) drop the old tables if they exist and write the new tables,
iii) check that all the data columns are contained in the main metadata table
iv) add datasetid columns and foreign key referencing datasets table, v) exit connection
}
\keyword{database}
