% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{writeValuesToDatabase}
\alias{writeValuesToDatabase}
\title{Function for writing a tibble as a table the database}
\usage{
writeValuesToDatabase(values, tableName, appendChoice, key = 0)
}
\arguments{
\item{values}{tibble to be added to the server}

\item{tableName}{name of table in the database}

\item{appendChoice}{TRUE to append to existing table of the same name; otherwise false}

\item{key}{column use to compare the old and new data; only new data unique in this column will be added (eg. use date or a key/id)}
}
\description{
This function will: i) connect to database, ii) write table; append if desired, iii) exit connection
}
\keyword{database}
