% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminfunctions.R
\name{writeNewDataset_single}
\alias{writeNewDataset_single}
\title{Function for writing a single new dataset from a csv to the database
This function will: i) connect to database, ii) drop the old table if exists and write the new table,
iii) add datasetid column and foreign key referencing datasets table, iv) exit connection}
\usage{
writeNewDataset_single(
  data_path,
  datasetid,
  base_path = "/Volumes/daten_inst/oekophysiologie/"
)
}
\arguments{
\item{data_path}{Path from the base path data folder to the csv data file}

\item{datasetid}{ID for the dataset to be added (the dataset must already have been added to the datasets table!)}

\item{base_path}{Path from root to the data folder (normally set as the server: "/Volumes/daten_inst/oekophysiologie/" but you must be connected to the server for this to work)}
}
\description{
Function for writing a single new dataset from a csv to the database
This function will: i) connect to database, ii) drop the old table if exists and write the new table,
iii) add datasetid column and foreign key referencing datasets table, iv) exit connection
}
\examples{
New example, test online
}
\keyword{database}
